+++ src/legacy/protection.go
@@ -1,8 +1,8 @@
 package legacy
 
 func Nox_xxx_protectionCreateStructForInt_56F280(a1 int, a2 int) int {
-	return int(nox_xxx_protectionCreateStructForInt_56F280(int32(a1), int32(a2)))
+	return int(nox_xxx_protectionCreateStructForInt_56F280(uint32(int32(a1)), int32(a2)))
 }
 func Nox_xxx_protectionCreateStructForFloat_56F480(a1 int, a2 float32) int {
-	return int(nox_xxx_protectionCreateStructForFloat_56F480(int32(a1), a2))
+	return int(nox_xxx_protectionCreateStructForFloat_56F480(uint32(int32(a1)), a2))
 }
